Solved day 13 part 2

+++ src/main/java/Day13ClawContraption/Point.java
@@ -1,3 +1,3 @@
 package Day13ClawContraption;
 
-public record Point(int x, int y) {}
+public record Point(long x, long y) {}
